% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detectors.R
\name{create_detector}
\alias{create_detector}
\title{Create a detection function}
\usage{
create_detector(FUN, ...)
}
\arguments{
\item{FUN}{a probability density (or mass) function whose first two arguments
must be \code{x} and \code{y}.}
}
\value{
a \code{function(x, y, ...)} where \code{x} and {y} are positions of
population specimens within a segment. The \code{FUN} is used to compute
the probabity that a specimens is detected, that is (\eqn{Pr(D|x,y;\theta)}),
\eqn{x,y} are a specimen's position and \eqn{\theta} are additional
parameters for the pdf (\code{FUN}). This function returns a \code{0/1}
vector of length \code{length(x)} (or \code{length(y)} if \code{x} is
missing), indicating whether specimens were not detected (1) or not (0).
}
\description{
Create a detection function
}
