% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/populations.R
\name{create_population_simulator}
\alias{create_population_simulator}
\title{Create a function for simulating a single species}
\usage{
create_population_simulator(FUN, ..., trait_map)
}
\arguments{
\item{FUN}{a data generating function}

\item{...}{arguments to curry out}

\item{trait_map}{a named character vector whose names are names of arguments
in \code{FUN} and whose elements are names of \link{species_traits}.}
}
\value{
a \code{function(traits, ...)} where \code{traits} is a list of species
traits and \code{...} are passed as  additional arguments to \code{FUN}.
}
\description{
Create a function for simulating a single species
}
\examples{
\dontrun{
  g <- create_generator(rMatClust, win = owin(c(0, 1), c(0, .25)),
                        trait_map = matern_traitmap)
  segments <- g(traits = speciesA$traits, scale = 0.05, nsim = 100)
}
}
