% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detectors.R
\name{detection_funs}
\alias{detection_funs}
\alias{oracle_detector}
\alias{half_norm_detector}
\title{Detection Functions}
\usage{
oracle_detector(x = NULL, y = NULL)

half_norm_detector(x = NULL, y = NULL)
}
\description{
Detects specimens with probability 1

Detects specimens with decaying probability with > y, irregardless of x.
}
