% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/space.R
\docType{data}
\name{example_study_area}
\alias{example_study_area}
\title{An example study area with a single property}
\format{An object of class \code{list} of length 1001.}
\usage{
example_study_area
}
\description{
An example study area with a single property
}
\keyword{datasets}
