% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/populations.R
\name{generate_matern_population}
\alias{generate_matern_population}
\title{Generate a single species population from a Matern Cluster Process}
\usage{
generate_matern_population(traits, area, ...)
}
\description{
Generates a population in a \code{\link{standard_segment}} using
\code{\link[spatstat]{rMatClust}}.
}
