% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/populations.R
\name{generate_matern_populations}
\alias{generate_matern_populations}
\title{Generate a multiple species populations from a Matern Cluster Process}
\usage{
generate_matern_populations(area, species_list, ...)
}
\description{
Generates mulitple populations in a \code{\link{standard_segment}} using
\code{\link[spatstat]{rMatClust}}.
}
