% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/populations.R
\docType{data}
\name{matern_traitmap}
\alias{matern_traitmap}
\title{Matern trait mapping}
\format{An object of class \code{character} of length 2.}
\usage{
matern_traitmap
}
\description{
Matern trait mapping
}
\keyword{datasets}
