% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/observers.R
\name{observers}
\alias{observers}
\alias{creater_observer}
\title{Create an observer}
\usage{
creater_observer(protocol, detectionFUN, detectionARGs = list())
}
\arguments{
\item{detectionFUN}{a \link{detection_funs}}

\item{detectionARGS}{a \code{list} of arguments passed to \code{detectionFUN}}
}
\description{
An observer is a function that takes a \code{\link{segment}} and
returns observations -- indicators of whether a population unit was detected.
An observer is defined by both a \link{detection_funs} and a set of fixed
arguments for a given detection function.
}
