% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/protocols.R
\name{protocols}
\alias{protocols}
\alias{cover_segment}
\alias{systematic}
\title{Protocols}
\usage{
cover_segment(segment)

systematic(segment)
}
\arguments{
\item{segment}{a study segment}
}
\description{
A protocol is a function that takes a segment and returns a list of (x, y)
pairs from which \code{\link{observers}} will use their \code{\link{detectors}}.
That is, a protocol defines the path by which \code{\link{observers}} traverse
a segment and take observations.
}
\section{Functions}{
\itemize{
\item \code{cover_segment}: observe the entire segment

\item \code{systematic}: observe 5 smaller quadrats within a segment
}}

