% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/species.R
\docType{data}
\name{species}
\alias{species}
\alias{speciesA}
\alias{speciesB}
\alias{speciesC}
\title{Species}
\format{An object of class \code{list} of length 3.}
\usage{
speciesA

speciesB

speciesC
}
\description{
In the context of this package, a \code{species} is an object with the following
elements:
}
\details{
\itemize{
\item name: the name of the species
\item aes: a \code{list} of aesthetic elements for plotting (e;g. color)
\item traits: a named \code{list} of functions that are used by \link{segment_generators}
to produce populations of species within a segment. See \link{species_traits}
for available traits
}
}
\keyword{datasets}
