% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/species.R
\name{species_traits}
\alias{species_traits}
\alias{create_locally_uniform_rarity_trait}
\alias{create_locally_uniform_clumping_trait}
\title{Trait generators}
\usage{
create_locally_uniform_rarity_trait(kappa, property_map)

create_locally_uniform_clumping_trait(mu, property_map)
}
\description{
Trait generators
}
