% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/space.R
\docType{data}
\name{standard_segment}
\alias{standard_segment}
\title{Study standard observation segment}
\format{An object of class \code{owin} of length 4.}
\usage{
standard_segment
}
\description{
The standard road segment size for this study in relative dimensions is
a 4:1 rectangle. See \code{\link[spatstat]{owin}}. The origin
(\code{(x = 0, y = 0)}) is the bottom-left corner of the shape (this is where
\link{observers} should start.
}
\keyword{datasets}
